[controller] Merge VeniceControllerConfig and VeniceControllerClusterConfig (#1070)
Today, in controllers, we have two classes of configs which serve the same purpose, and it is unnecessary to maintain both. This commit merges "VeniceControllerConfig" and "VeniceControllerClusterConfig".
