+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/controller/AbstractTestVeniceHelixAdmin.java
@@ -68,7 +68,7 @@ class AbstractTestVeniceHelixAdmin {
   VeniceHelixAdmin veniceAdmin;
   String clusterName;
   String storeOwner = "Doge of Venice";
-  VeniceControllerConfig controllerConfig;
+  VeniceControllerClusterConfig controllerConfig;
   String zkAddress;
   ZkServerWrapper zkServerWrapper;
   PubSubBrokerWrapper pubSubBrokerWrapper;
@@ -104,7 +104,7 @@ class AbstractTestVeniceHelixAdmin {
     properties.put(UNREGISTER_METRIC_FOR_DELETED_STORE_ENABLED, true);
     controllerProps = new VeniceProperties(properties);
     helixMessageChannelStats = new HelixMessageChannelStats(new MetricsRepository(), clusterName);
-    controllerConfig = new VeniceControllerConfig(controllerProps);
+    controllerConfig = new VeniceControllerClusterConfig(controllerProps);
     multiClusterConfig = TestUtils.getMultiClusterConfigFromOneCluster(controllerConfig);
     veniceAdmin = new VeniceHelixAdmin(
         multiClusterConfig,
