+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/controller/TestVeniceHelixResources.java
@@ -57,7 +57,7 @@ public class TestVeniceHelixResources {
         .getReadOnlyZKSharedSystemStoreRepository();
     doReturn(mock(HelixReadOnlyZKSharedSchemaRepository.class)).when(veniceHelixAdmin)
         .getReadOnlyZKSharedSchemaRepository();
-    VeniceControllerConfig controllerConfig = mock(VeniceControllerConfig.class);
+    VeniceControllerClusterConfig controllerConfig = mock(VeniceControllerClusterConfig.class);
     when(controllerConfig.getDaVinciPushStatusScanThreadNumber()).thenReturn(4);
     when(controllerConfig.getDaVinciPushStatusScanIntervalInSeconds()).thenReturn(5);
     when(controllerConfig.isDaVinciPushStatusEnabled()).thenReturn(true);
@@ -67,7 +67,7 @@ public class TestVeniceHelixResources {
         zkClient,
         new HelixAdapterSerializer(),
         new SafeHelixManager(controller),
-        mock(VeniceControllerConfig.class),
+        mock(VeniceControllerClusterConfig.class),
         veniceHelixAdmin,
         metricsRepository,
         mock(RealTimeTopicSwitcher.class),
