+++ services/venice-controller/src/main/java/com/linkedin/venice/controller/StoreGraveyardCleanupService.java
@@ -65,7 +65,7 @@ public class StoreGraveyardCleanupService extends AbstractVeniceService {
           time.sleep((long) sleepIntervalBetweenListFetchMinutes * Time.MS_PER_MINUTE);
           // loop all clusters
           for (String clusterName: multiClusterConfig.getClusters()) {
-            VeniceControllerConfig clusterConfig = multiClusterConfig.getControllerConfig(clusterName);
+            VeniceControllerClusterConfig clusterConfig = multiClusterConfig.getControllerConfig(clusterName);
             boolean cleanupEnabled = clusterConfig.isStoreGraveyardCleanupEnabled();
             int delayInMinutes = clusterConfig.getStoreGraveyardCleanupDelayMinutes();
             if (!cleanupEnabled || !admin.isLeaderControllerFor(clusterName)) {
