+++ services/venice-controller/src/main/java/com/linkedin/venice/controller/VeniceController.java
@@ -258,7 +258,8 @@ public class VeniceController {
 
   private void initializeSystemSchema(Admin admin) {
     String systemStoreCluster = multiClusterConfigs.getSystemSchemaClusterName();
-    VeniceControllerConfig systemStoreClusterConfig = multiClusterConfigs.getControllerConfig(systemStoreCluster);
+    VeniceControllerClusterConfig systemStoreClusterConfig =
+        multiClusterConfigs.getControllerConfig(systemStoreCluster);
     if (!multiClusterConfigs.isParent() && systemStoreClusterConfig.isSystemSchemaInitializationAtStartTimeEnabled()) {
       String regionName = systemStoreClusterConfig.getRegionName();
       String childControllerUrl = systemStoreClusterConfig.getChildControllerUrl(regionName);
