+++ services/venice-controller/src/test/java/com/linkedin/venice/controller/AbstractTestVeniceParentHelixAdmin.java
@@ -65,7 +65,7 @@ public class AbstractTestVeniceParentHelixAdmin {
 
   TopicManager topicManager;
   VeniceHelixAdmin internalAdmin;
-  VeniceControllerConfig config;
+  VeniceControllerClusterConfig config;
   ZkClient zkClient;
   VeniceWriter veniceWriter;
   VeniceParentHelixAdmin parentAdmin = null;
@@ -186,8 +186,8 @@ public class AbstractTestVeniceParentHelixAdmin {
     }
   }
 
-  VeniceControllerConfig mockConfig(String clusterName) {
-    VeniceControllerConfig config = mock(VeniceControllerConfig.class);
+  VeniceControllerClusterConfig mockConfig(String clusterName) {
+    VeniceControllerClusterConfig config = mock(VeniceControllerClusterConfig.class);
     doReturn(clusterName).when(config).getClusterName();
     doReturn(KAFKA_REPLICA_FACTOR).when(config).getKafkaReplicationFactor();
     doReturn(KAFKA_REPLICA_FACTOR).when(config).getAdminTopicReplicationFactor();
@@ -207,7 +207,7 @@ public class AbstractTestVeniceParentHelixAdmin {
     return config;
   }
 
-  HelixVeniceClusterResources mockResources(VeniceControllerConfig config, String clusterName) {
+  HelixVeniceClusterResources mockResources(VeniceControllerClusterConfig config, String clusterName) {
     HelixVeniceClusterResources resources = mock(HelixVeniceClusterResources.class);
     doReturn(config).when(resources).getConfig();
     doReturn(resources).when(internalAdmin).getHelixVeniceClusterResources(clusterName);
