+++ services/venice-controller/src/test/java/com/linkedin/venice/controller/TestVeniceControllerClusterConfig.java
@@ -12,7 +12,7 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 
-public class TestVeniceControllerConfig {
+public class TestVeniceControllerClusterConfig {
   private static final String DELIMITER = ",\\s*";
   private static final String WHITE_LIST = "dc1,dc2";
 
@@ -22,7 +22,7 @@ public class TestVeniceControllerConfig {
     builder.put("child.cluster.url.dc1", "http://host:1234, http://host:5678")
         .put("child.cluster.url.dc2", "http://host:1234, http://host:5678");
 
-    Map<String, String> map = VeniceControllerConfig.parseClusterMap(builder.build(), WHITE_LIST);
+    Map<String, String> map = VeniceControllerClusterConfig.parseClusterMap(builder.build(), WHITE_LIST);
 
     Assert.assertEquals(map.size(), 2);
     Assert.assertTrue(map.keySet().contains("dc1"));
@@ -38,7 +38,7 @@ public class TestVeniceControllerConfig {
     PropertyBuilder builder = new PropertyBuilder();
     builder.put("child.cluster.d2.zkHost.dc1", "zkAddress1").put("child.cluster.d2.zkHost.dc2", "zkAddress2");
 
-    Map<String, String> map = VeniceControllerConfig.parseClusterMap(builder.build(), WHITE_LIST, true);
+    Map<String, String> map = VeniceControllerClusterConfig.parseClusterMap(builder.build(), WHITE_LIST, true);
     Assert.assertEquals(map.get("dc1").split(DELIMITER).length, 1);
     Assert.assertTrue(map.get("dc2").split(DELIMITER)[0].equals("zkAddress2"));
   }
@@ -51,7 +51,7 @@ public class TestVeniceControllerConfig {
 
     // Add the list of disabled endpoints, '/' are optional, and will be ignored. Invalid values will be filtered
     builder.put(CONTROLLER_DISABLED_ROUTES, "request_topic, /discover_cluster, foo,bar");
-    List<ControllerRoute> parsedRoutes = VeniceControllerConfig
+    List<ControllerRoute> parsedRoutes = VeniceControllerClusterConfig
         .parseControllerRoutes(builder.build(), CONTROLLER_DISABLED_ROUTES, Collections.emptyList());
 
     // Make sure it looks right.
@@ -64,27 +64,27 @@ public class TestVeniceControllerConfig {
   public void emptyAllowlist() {
     PropertyBuilder build = new PropertyBuilder().put("child.cluster.url.dc1", "http://host:1234, http://host:5678")
         .put("child.cluster.url.dc2", "http://host:1234, http://host:5678");
-    VeniceControllerConfig.parseClusterMap(build.build(), "");
+    VeniceControllerClusterConfig.parseClusterMap(build.build(), "");
   }
 
   @Test(expectedExceptions = VeniceException.class)
   public void nullAllowlist() {
     PropertyBuilder build = new PropertyBuilder().put("child.cluster.url.dc1", "http://host:1234, http://host:5678")
         .put("child.cluster.url.dc2", "http://host:1234, http://host:5678");
-    VeniceControllerConfig.parseClusterMap(build.build(), "");
+    VeniceControllerClusterConfig.parseClusterMap(build.build(), "");
   }
 
   @Test(expectedExceptions = VeniceException.class)
   public void errOnMissingScheme() {
     PropertyBuilder builder = new PropertyBuilder();
     builder.put("child.cluster.url.dc1", "host:1234");
-    VeniceControllerConfig.parseClusterMap(builder.build(), WHITE_LIST);
+    VeniceControllerClusterConfig.parseClusterMap(builder.build(), WHITE_LIST);
   }
 
   @Test(expectedExceptions = VeniceException.class)
   public void errOnMissingNodes() {
     PropertyBuilder builder = new PropertyBuilder();
     builder.put("child.cluster.url.dc1", "");
-    VeniceControllerConfig.parseClusterMap(builder.build(), WHITE_LIST);
+    VeniceControllerClusterConfig.parseClusterMap(builder.build(), WHITE_LIST);
   }
 }
